+++ changelog.md
@@ -5,6 +5,7 @@
 * Added experimental support to "subscribers" buffer, inside frienship requests. This shows friend requests that have been declined by the current user.
 * Fixed an error in Socializer that was making it unable to detect unread messages properly.
 * Socializer should save all tracebacks directly to error.log instead of displaying an error message during exit. ([#27,](https://code.manuelcortez.net/manuelcortez/socializer/issues/27))
+* When displaying user profiles, fixed an error where a married person without specifing relation partner would cause an error in the program. ([#29,](https://code.manuelcortez.net/manuelcortez/socializer/issues/29))
 
 ## Changes in version 0.18 (21.01.2019)